Restructure the MATLAB server and the client for the remote access. (NOT QUITE FINISHED MatlabCode needs some revision)
The way the client was interacting with the proxy on the remote server did not work at all, since it is not possible to get an object from the remote JVM and then manipulate it as if the client would “be” on the remote machine (see cajo docs)
So now the MatlabServer also implements the MatlabOperations interface from the MatlabControl library. This is to forward all the calls to MATLAB to the server and then get back static content.
The development of the task-methods in of the MatlabClient-Remote is complete but not yet tested.

The file manipulations in the MatlabCode class were removed. The method function signatures have become a bit more complex, but on the other hand what they do has become more transparent. Since this class has outgrown itself this re-organisation has become necessary. Some of the manipulations are done now in the MatlabFileTransfer class, which has added functionality to accommodate transfer of data for processing on the local machine as well as the remote host.

+++ de.mpicbg.knime.scripting.matlab.srv/src/de/mpicbg/knime/scripting/matlab/srv/utils/MatlabFileTransfer.java
@@ -1,20 +1,37 @@
 package de.mpicbg.knime.scripting.matlab.srv.utils;
 
-import java.io.*;
+
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.Arrays;
 
+import org.apache.commons.io.FilenameUtils;
+
+import de.mpicbg.knime.scripting.matlab.srv.Matlab;
 import de.mpicbg.knime.scripting.matlab.srv.MatlabRemote;
 
 
 /**
- * Encapsulates client and server temp-files to simplify create/delete/copy operations
+ * Encapsulates client and server temp-files to simplify create/delete/copy operations.
+ * Since this wrapper can be used for local and remote MATLAB hosts, the temp-files
+ * are created only upon request (getter methods).
  * 
- * @author Tom Haux
+ * @author Tom Haux, Felix Meyenhofer
  */
 public class MatlabFileTransfer {
 
 	/** The files on client and server side */
 	private File clientFile, serverFile;
+	
+	/** Prefix and suffix of the temp-files */
+	private String prefix, suffix;
 	
 	/** Object to communicate with the server */
 	private MatlabRemote server;
@@ -22,24 +39,26 @@ public class MatlabFileTransfer {
 	
 	/**
 	 * Constructor for a transfer file (that exists on the server and the client side)
-	 * in the JVM temp-directory by defining file prefix and suffix.
+	 * in the JVM temp-directory by defining file prefix and suffix. Both files will
+	 * just be initialized.
 	 * 
-	 * @param matlab MATLAB server object
+	 * @param matlab {@link MatlabRemote} server object
 	 * @param prefix Temp-file prefix
 	 * @param suffix Temp-file suffix
 	 * @throws IOException
 	 */
 	public MatlabFileTransfer(MatlabRemote matlab, String prefix, String suffix) throws IOException {
 		this.server = matlab;
-
-		clientFile = File.createTempFile(prefix, suffix);
-		clientFile.deleteOnExit();
-		serverFile = matlab.createTempFile(prefix, suffix);
+		this.prefix = prefix;
+		this.suffix = suffix;
 	}
 
 	/**
-	 * Constructor for a transfer file (that exists on the server and the client side)
-	 * in the JVM temp-directory by giving in the local {@link File} object.
+	 * Constructor for a transfer file that will be synchronized from the client to
+	 * server immediately.
+	 * This is deduced by the fact that we have a MATLAB server, and an existing file
+	 * (that was created before presumably with content) so the next logical step is to
+	 * make it available on the server. 
 	 * 
 	 * @param matlab MATLAB server object
 	 * @param inputFile	Local temp-file
@@ -47,10 +66,53 @@ public class MatlabFileTransfer {
 	 */
 	public MatlabFileTransfer(MatlabRemote matlab, File inputFile) throws IOException {
 		this.server = matlab;
-
-		clientFile = inputFile;
-		serverFile = matlab.createTempFile("MatlabRemoteServer", ".tmp");
+		this.prefix = FilenameUtils.getBaseName(inputFile.getAbsolutePath());
+		this.suffix = FilenameUtils.getExtension(inputFile.getAbsolutePath());
+		this.clientFile = inputFile;
+		
+		upload();
 	}
+	
+	/**
+	 * Constructor for a file in the local mode. We don't have a MATLAB server object
+	 * 
+	 * @param inputFile
+	 */
+	public MatlabFileTransfer(File inputFile) {
+		this.clientFile = inputFile;
+	}
+	
+	
+	/**
+	 * Constructor for a transfer file, where the resource from this package will 
+	 * be streamed immediately to the server temp-file. 
+	 * 
+	 * @param matlab
+	 * @param resourcePath
+	 * @throws FileNotFoundException
+	 * @throws IOException
+	 */
+	public MatlabFileTransfer(MatlabRemote matlab, String resourcePath) throws FileNotFoundException, IOException {
+		this(resourcePath);
+		upload();
+	}
+	
+	/**
+	 * Constructor for a transfer file where the resource will be streamed immediately
+	 * to the local JVM temp-directory to be made available for MATLAB.
+	 * 
+	 * @param resourcePath
+	 * @throws FileNotFoundException
+	 * @throws IOException
+	 */
+	public MatlabFileTransfer(String resourcePath) throws FileNotFoundException, IOException {
+		clientFile = new File(Matlab.TEMP_PATH, FilenameUtils.getName(resourcePath));
+		clientFile.deleteOnExit();
+        ClassLoader loader = Thread.currentThread().getContextClassLoader();
+        InputStream resstream = loader.getResourceAsStream(resourcePath);
+        writeStreamToFile(resstream, new FileOutputStream(clientFile));
+	}
+
 
 	/**
 	 * Get the client file object on the local machine.
@@ -58,6 +120,16 @@ public class MatlabFileTransfer {
 	 * @return Client file on local machine
 	 */
 	public File getClientFile() {
+		if (clientFile == null) {
+			try {
+				clientFile = File.createTempFile(prefix, suffix);
+			} catch (IOException e) {
+				e.printStackTrace();
+				throw new RuntimeException("MATLAB file transfer: unable to create local file (client side).");
+			}
+			clientFile.deleteOnExit();
+		}
+			
 		return clientFile;
 	}
 
@@ -66,7 +138,15 @@ public class MatlabFileTransfer {
 	 * 
 	 * @return Temp-file object of the server (on remote machine)
 	 */
-	public File getServerFile() {
+	private File getServerFile() {
+		if (serverFile == null) {
+			try {
+				serverFile = server.createTempFile(prefix, suffix);
+			} catch (IOException e) {
+				e.printStackTrace();
+				throw new RuntimeException("MATLAB file transfer: unable to create remote file (server side).");
+			}
+		}
 		return serverFile;
 	}
 
@@ -76,7 +156,7 @@ public class MatlabFileTransfer {
 	 * @return Relative path to the local temp-file
 	 */
 	public String getClientPath() {
-		return clientFile.getAbsolutePath();
+		return getClientFile().getAbsolutePath();
 	}
 
 	/**
@@ -85,21 +165,21 @@ public class MatlabFileTransfer {
 	 * @return Relative path on the remote machine
 	 */
 	public String getServerPath() {
-		return server.getFilePath(serverFile);
+		return getServerFile().getAbsolutePath();
 	}
 
 	/**
 	 * Upload the content from the local file (client) to the remote file (server)
 	 */
-	public void upload() {
+	public void upload() { // TODO remove redundant code (I left it so fare to be sure we still have a working version)
 		// Read the file into a byte array and pass it to the server
 		int descriptor = -1;
 		BufferedInputStream bis = null;
 
 		try {
 			byte[] buffer = new byte[8192];
-			descriptor = server.openFile(serverFile);
-			bis = new BufferedInputStream(new FileInputStream(clientFile));
+			descriptor = server.openFile(getServerFile());
+			bis = new BufferedInputStream(new FileInputStream(getClientFile()));
 			int bytesRead = 0;
 
 			// Keep reading until we hit the end of the file
@@ -123,20 +203,49 @@ public class MatlabFileTransfer {
 			}
 		}
 	}
+	
+	
+	public void upload(InputStream bis) throws IOException {
+		// Read the file into a byte array and pass it to the server
+		int descriptor = -1;
+
+
+		try {
+			byte[] buffer = new byte[8192];
+			descriptor = server.openFile(getServerFile());
+			bis = new BufferedInputStream(bis);
+			int bytesRead = 0;
+
+			// Keep reading until we hit the end of the file
+			// when the end of the stream has been reached, -1 is returned
+			while ((bytesRead = bis.read(buffer)) != -1) {
+				byte[] b = buffer;
+				if (bytesRead < buffer.length) b = Arrays.copyOf(buffer, bytesRead);
+				server.writeFile(descriptor, b);
+			}
+		} catch (IOException e) {
+			throw new RuntimeException("MATLAB file transfer: the upload was aborted.");
+		} finally {
+			if (descriptor != -1) 
+				server.closeFile(descriptor);
+			if (bis != null)
+				bis.close();
+		}
+	}
 
 	/**
 	 * Fetch/download the content from the remote file (server) to the local file (server).
 	 */
-	public void fetch() {
+	public void download() {
 		// Get a byte array from the server and write it to a file
 		// Read the file into a byte array and pass it to the server
 		int descriptor = -1;
 		BufferedOutputStream bos = null;
 
 		try {
-			descriptor = server.openFile(serverFile);
+			descriptor = server.openFile(getServerFile());
 
-			bos = new BufferedOutputStream(new FileOutputStream(clientFile));
+			bos = new BufferedOutputStream(new FileOutputStream(getClientFile()));
 			while (true) {
 				byte[] bytes = server.readFile(descriptor);
 				if (bytes.length == 0) break;
@@ -161,7 +270,31 @@ public class MatlabFileTransfer {
 	 * Delete the transfer temp-file (on local and remote machine)
 	 */
 	public void delete() {
-		clientFile.delete();
-		server.deleteFile(serverFile);
+		if (clientFile != null)
+			clientFile.delete();
+		if (serverFile != null)
+			server.deleteFile(serverFile);
 	}
+	
+	
+	 /**
+     * Write an file input to an output stream.
+     * 
+     * @param in
+     * @param out
+     * @throws IOException
+     */
+    private static void writeStreamToFile(InputStream in, OutputStream out) throws IOException {
+        byte[] buffer = new byte[16384];
+        while (true) {
+            int count = in.read(buffer);
+            if (count < 0)
+                break;
+            out.write(buffer, 0, count);
+        }
+        in.close();
+        out.close();
+    }
+
+	
 }
